ToolsMCP/requirements.txt
fastmcp==2.12.3
langchain==0.3.27
langchain_core==0.3.76
langchain_google_genai==2.1.12
langchain_openai==0.3.33
pydantic==2.11.9
python-dotenv==1.1.1
Requests==2.32.5
sentence_transformers==5.1.0
torch==2.8.0
